Allow any integer type
Realistically, this will raise an exception if you supply an int larger
than `Int64::MAX`, but if you're keeping 18 quintillion values per key,
you're probably not going to achieve the results you're aiming for.

+++ src/kv.cr
@@ -282,7 +282,7 @@ module NATS
         description : String = "",
         *,
         max_value_size : Int32? = nil,
-        history : UInt8 = 1,
+        history : Int = 1,
         ttl : Time::Span? = nil,
         max_bytes : Int64? = nil,
         storage : JetStream::StreamConfig::Storage = :file,
